pyproject.toml
[tool.poetry]
name = "nonebot_plugin_ddcheck"
version = "0.4.3"
description = "Nonebot2 成分姬插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-ddcheck"
repository = "https://github.com/noneplugin/nonebot-plugin-ddcheck"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-plugin-alconna = ">=0.46.3,<1.0.0"
nonebot-plugin-htmlrender = "^0.4.0"
nonebot-plugin-apscheduler = ">=0.4.0,<1.0.0"
nonebot-plugin-localstore = ">=0.6.0,<1.0.0"
httpx = ">=0.20.0,<1.0.0"
Jinja2 = "^3.0.0"

[tool.poetry.group.dev.dependencies]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
